requirements.txt
tensorflow[and-cuda]
tf-keras
matplotlib
scipy
numpy
pandas
requests
scikit-learn
pydot
transformers
boto3